refactor(models,ui): share provider alias resolution
Extract PROVIDER_ALIASES + resolveProviderName from registry.ts into a
new browser-safe module and expose it from both index.ts and browser.ts.

ProviderIcon now resolves the canonical provider name before the icon
lookup, so aliased providers like openai-codex render the OpenAI icon
without needing per-variant entries in PROVIDER_ICON_MAP. The previous
direct openai-codex entry is reverted in favor of this shared path.
