+++ packages/ui/package.json
@@ -35,6 +35,7 @@
     "@radix-ui/react-switch": "^1.2.6",
     "@radix-ui/react-toast": "^1.2.15",
     "@radix-ui/react-tooltip": "^1.2.8",
+    "@domain/models": "workspace:*",
     "@repo/utils": "workspace:*",
     "@tanstack/react-virtual": "^3.13.23",
     "class-variance-authority": "^0.7.1",
