+++ packages/ui/src/components/icons/custom-icons/providers/provider-icon.tsx
@@ -1,3 +1,4 @@
+import { resolveProviderName } from "@domain/models"
 import { BotIcon } from "lucide-react"
 import { Icon, type IconProps } from "../../icons.tsx"
 import { PROVIDER_ICON_MAP } from "./provider-map.ts"
@@ -7,7 +8,8 @@ export interface ProviderIconProps extends Omit<IconProps, "icon"> {
 }
 
 export function ProviderIcon({ provider, ...props }: ProviderIconProps) {
-  const supportedIcon = PROVIDER_ICON_MAP[provider.toLowerCase()]
+  const resolved = resolveProviderName(provider).toLowerCase()
+  const supportedIcon = PROVIDER_ICON_MAP[resolved]
   const IconComponent = supportedIcon ?? BotIcon
   return <Icon icon={IconComponent} {...props} />
 }
